add more events endpoint

+++ ewm-main-service/src/main/java/ru/practicum/ewm/dto/events/EventShortDto.java
@@ -23,7 +23,7 @@ public class EventShortDto {
     @NotBlank
     private CategoryDto category;
 
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
 
     @NotBlank
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
